+++ examples/cloud-grpc-client/src/main/java/net/devh/examples/grpc/cloud/GrpcClientController.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.cloud;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
