+++ examples/cloud-grpc-client/src/main/java/net/devh/examples/grpc/cloud/GrpcClientService.java
@@ -1,13 +1,13 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.cloud;
 
+import net.devh.examples.grpc.lib.HelloReply;
+import net.devh.examples.grpc.lib.HelloRequest;
+import net.devh.examples.grpc.lib.SimpleGrpc;
 import net.devh.springboot.autoconfigure.grpc.client.GrpcClient;
 
 import org.springframework.stereotype.Service;
 
 import io.grpc.Channel;
-import io.grpc.examples.helloworld.GreeterGrpc;
-import io.grpc.examples.helloworld.HelloReply;
-import io.grpc.examples.helloworld.HelloRequest;
 
 /**
  * User: Michael
@@ -17,11 +17,11 @@ import io.grpc.examples.helloworld.HelloRequest;
 @Service
 public class GrpcClientService {
 
-    @GrpcClient("local-grpc-server")
+    @GrpcClient("cloud-grpc-server")
     private Channel serverChannel;
 
     public String sendMessage(String name) {
-        GreeterGrpc.GreeterBlockingStub stub = GreeterGrpc.newBlockingStub(serverChannel);
+        SimpleGrpc.SimpleBlockingStub stub = SimpleGrpc.newBlockingStub(serverChannel);
         HelloReply response = stub.sayHello(HelloRequest.newBuilder().setName(name).build());
         return response.getMessage();
     }
