+++ examples/cloud-grpc-client/src/main/java/net/devh/examples/grpc/cloud/LogGrpcInterceptor.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.cloud;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
