+++ examples/grpc-lib/src/main/proto/helloworld.proto
@@ -1,14 +1,14 @@
 syntax = "proto3";
 
 option java_multiple_files = true;
-option java_package = "io.grpc.examples.helloworld";
+option java_package = "net.devh.examples.grpc.lib";
 option java_outer_classname = "HelloWorldProto";
 option objc_class_prefix = "HLW";
 
 package helloworld;
 
 // The greeting service definition.
-service Greeter {
+service Simple {
     // Sends a greeting
     rpc SayHello (HelloRequest) returns (HelloReply) {
     }
