+++ examples/local-grpc-client/src/main/java/net/devh/examples/grpc/local/GlobalClientInterceptorConfiguration.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.local;
 
 import net.devh.springboot.autoconfigure.grpc.client.GlobalClientInterceptorConfigurerAdapter;
 import net.devh.springboot.autoconfigure.grpc.client.GlobalClientInterceptorRegistry;
