+++ examples/local-grpc-client/src/main/java/net/devh/examples/grpc/local/GrpcClientApplication.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.local;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,9 +9,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * Date: 2016/11/8
  */
 @SpringBootApplication
-public class GrpcServerApplication {
+public class GrpcClientApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(GrpcServerApplication.class, args);
+        SpringApplication.run(GrpcClientApplication.class, args);
     }
 }
