+++ examples/local-grpc-server/src/main/java/net/devh/examples/grpc/local/GlobalInterceptorConfiguration.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.local;
 
 import net.devh.springboot.autoconfigure.grpc.server.GlobalServerInterceptorConfigurerAdapter;
 import net.devh.springboot.autoconfigure.grpc.server.GlobalServerInterceptorRegistry;
