+++ examples/local-grpc-server/src/main/java/net/devh/examples/grpc/local/GrpcServerApplication.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.local;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,9 +9,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * Date: 2016/11/8
  */
 @SpringBootApplication
-public class GrpcClientApplication {
+public class GrpcServerApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(GrpcClientApplication.class, args);
+        SpringApplication.run(GrpcServerApplication.class, args);
     }
 }
