+++ examples/local-grpc-server/src/main/java/net/devh/examples/grpc/local/LogGrpcInterceptor.java
@@ -1,4 +1,4 @@
-package net.devh.grpc;
+package net.devh.examples.grpc.local;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
